relativedelta.months only returns the difference between the month numbers, eg. it returns 4 for the difference between May 2010 and Sept 2011, month 5 to month 9.
So in models.py, we've refactored month difference calculation into it's own method, taking into account the number of years by which two given dates are separated.

In tests.py, we simply perform the same calculation as in Contract.__month_diff manually in the test to arrive at the same result.

Also, we import from dateutil.relativedelta, instead of one level higher with just dateutil